Add files via upload
This is the experimental data for the mixed gases in air flow, the responses represent a sensor discrete response signal. A data collecting system was used to measure the current-voltage characteristics of sensor responses under repeated gas flows (500 sccm) of Nitrogen Oxide (diluted in standard air) and standard air (O2: 20%, N2: 80%) controlled by a mass flow controller. The sample was kept at 180°C in a temperature-controlled electric furnace. It was discovered that when the adsorption is at equilibrium, the discrete signal from mixed gases displays signal additivity and that the curves behave linearly with time. The flowing gas causes this phenomenon.